Create ALTERNATIVE_ISSUE_TEMPLATE.yaml
This file is added to check the issue form functionality that GitHub provides as an alternative to a regular issue template (https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository#creating-issue-forms). 
Issue form has validations and in general looks better that a plain text. We need to add this file to be able to test it. Without having it it's impossible to test even locally.